+++ src/CSSList/Document.php
@@ -8,7 +8,6 @@ use Sabberworm\CSS\Parsing\SourceException;
 use Sabberworm\CSS\Property\Selector;
 use Sabberworm\CSS\RuleSet\DeclarationBlock;
 use Sabberworm\CSS\RuleSet\RuleSet;
-use Sabberworm\CSS\Value\Value;
 
 /**
  * This class represents the root of a parsed CSS file. It contains all top-level CSS contents: mostly declaration
@@ -74,33 +73,6 @@ class Document extends CSSBlockList
         /** @var array<int, RuleSet> $aResult */
         $aResult = [];
         $this->allRuleSets($aResult);
-        return $aResult;
-    }
-
-    /**
-     * Returns all `Value` objects found recursively in `Rule`s in the tree.
-     *
-     * @param CSSList|RuleSet|string $mElement
-     *        the `CSSList` or `RuleSet` to start the search from (defaults to the whole document).
-     *        If a string is given, it is used as rule name filter.
-     * @param bool $bSearchInFunctionArguments whether to also return Value objects used as Function arguments.
-     *
-     * @return array<int, Value>
-     *
-     * @see RuleSet->getRules()
-     */
-    public function getAllValues($mElement = null, $bSearchInFunctionArguments = false)
-    {
-        $sSearchString = null;
-        if ($mElement === null) {
-            $mElement = $this;
-        } elseif (is_string($mElement)) {
-            $sSearchString = $mElement;
-            $mElement = $this;
-        }
-        /** @var array<int, Value> $aResult */
-        $aResult = [];
-        $this->allValues($mElement, $aResult, $sSearchString, $bSearchInFunctionArguments);
         return $aResult;
     }